レンダラの修正

+++ src/renderer/model_matrix.rs
@@ -2,7 +2,7 @@ use ndarray::{arr2, Array2};
 use std::convert::Into;
 
 pub struct ModelMatrix {
-    model_matrix: Array2<f64>,
+    model_matrix: Array2<f32>,
 }
 
 impl ModelMatrix {
@@ -17,7 +17,7 @@ impl ModelMatrix {
         }
     }
 
-    pub fn with_movement(mut self, movement: &[f64; 3]) -> Self {
+    pub fn with_movement(mut self, movement: &[f32; 3]) -> Self {
         let m = movement;
         let t = arr2(&[
             [1.0, 0.0, 0.0, 0.0],
@@ -29,7 +29,7 @@ impl ModelMatrix {
         self
     }
 
-    pub fn with_scale(mut self, scale: &[f64; 3]) -> Self {
+    pub fn with_scale(mut self, scale: &[f32; 3]) -> Self {
         let s = scale;
         let t = arr2(&[
             [s[0], 0.0, 0.0, 0.0],
@@ -41,7 +41,7 @@ impl ModelMatrix {
         self
     }
 
-    pub fn with_x_axis_rotation(mut self, x_axis_rotation: f64) -> Self {
+    pub fn with_x_axis_rotation(mut self, x_axis_rotation: f32) -> Self {
         let (s, c) = x_axis_rotation.sin_cos();
         let t = arr2(&[
             [1.0, 0.0, 0.0, 0.0],
@@ -54,7 +54,7 @@ impl ModelMatrix {
     }
 
     #[allow(dead_code)]
-    pub fn with_y_axis_rotation(mut self, y_axis_rotation: f64) -> Self {
+    pub fn with_y_axis_rotation(mut self, y_axis_rotation: f32) -> Self {
         let (s, c) = y_axis_rotation.sin_cos();
         let t = arr2(&[
             [c, 0.0, s, 0.0],
@@ -66,7 +66,7 @@ impl ModelMatrix {
         self
     }
 
-    pub fn with_z_axis_rotation(mut self, z_axis_rotation: f64) -> Self {
+    pub fn with_z_axis_rotation(mut self, z_axis_rotation: f32) -> Self {
         let (s, c) = z_axis_rotation.sin_cos();
         let t = arr2(&[
             [c, -s, 0.0, 0.0],
@@ -79,8 +79,8 @@ impl ModelMatrix {
     }
 }
 
-impl Into<Array2<f64>> for ModelMatrix {
-    fn into(self) -> Array2<f64> {
+impl Into<Array2<f32>> for ModelMatrix {
+    fn into(self) -> Array2<f32> {
         self.model_matrix
     }
 }
